requirements.txt
yfinance
discord.py